rename debugger_ui elements to variable_list
