+++ crates/debugger_ui/src/variable_list.rs
